+++ client/dashboard/sites/settings/index.tsx
@@ -8,6 +8,8 @@ import { __ } from '@wordpress/i18n';
 import { siteQuery, siteSettingsQuery } from '../../app/queries';
 import { PageHeader } from '../../components/page-header';
 import PageLayout from '../../components/page-layout';
+import DatabaseSettingsSummary from '../settings-database/summary';
+import PHPSettingsSummary from '../settings-php/summary';
 import SiteVisibilitySettingsSummary from '../settings-site-visibility/summary';
 import SubscriptionGiftingSettingsSummary from '../settings-subscription-gifting/summary';
 import WordPressSettingsSummary from '../settings-wordpress/summary';
@@ -34,7 +36,9 @@ export default function SiteSettings( { siteSlug }: { siteSlug: string } ) {
 			<Heading>{ __( 'Server' ) }</Heading>
 			<Card>
 				<VStack>
+					<DatabaseSettingsSummary site={ site } />
 					<WordPressSettingsSummary site={ site } />
+					<PHPSettingsSummary site={ site } />
 				</VStack>
 			</Card>
 			<SiteActions site={ site } />