+++ packages/components/.storybook/preview.js
@@ -12,7 +12,7 @@ const parameters = {
 	},
 	options: {
 		storySort: ( a, b ) => {
-			const sectionOrder = [ 'WP Overrides', 'Deprecated', 'Unaudited' ];
+			const sectionOrder = [ 'Validated Form Controls', 'WP Overrides', 'Deprecated', 'Unaudited' ];
 			const aIndex = sectionOrder.findIndex( ( prefix ) => a.title.startsWith( prefix ) );
 			const bIndex = sectionOrder.findIndex( ( prefix ) => b.title.startsWith( prefix ) );