+++ packages/components/CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Enhancements
 
 - Add `BigSkyLogo.Mark` component ([#103612](https://github.com/Automattic/wp-calypso/pull/103612)).
+- Add `ValidatedFormControls` components, still in beta ([#100771](https://github.com/Automattic/wp-calypso/pull/100771)).
 
 ## 2.3.0